(PC-30735) refactor(gradient): extract onViewableItemsChanged
